añadidos menus de users, sessions y game en app_menu

+++ app_menu.py
@@ -5,6 +5,7 @@ from os import system
 API_URL = 'http://localhost:8080'
 API_URL_GAMEMODE = API_URL + '/game/'
 API_URL_USER = API_URL + '/user/'
+API_URL_USER_GAME = API_URL_USER + '{}/game/{}'
 
 
 TEXT_LINE = '\n------------------------------------'
@@ -38,7 +39,7 @@ TEXT_MENU_MANAGE_USERS = """\n
 
 TEXT_MENU_MANAGE_GAMESESSIONS = """\n
     Opciones:
-    - 1: Ver partidas a las que te has apuntado
+    - 1: Cargar partida
     - 2: Apuntarse a una partida
     - 3: Quitarse de una partida
 
@@ -52,8 +53,8 @@ TEXT_MENU_PLAY_GAME = """\n
     - 3: Comprar ventaja
     - 4: Mostrar filtros activos
 
-    - 9: Elegir otra sesión de juego
-    - 0: Cerrar aplicación
+    - 9: Cerrar aplicación
+    - 0: Volver a la pantalla de usuario
 """
 TEXT_MENU_PLAY_GAME = TEXT_LINE + TEXT_MENU_PLAY_GAME + TEXT_LINE
 
@@ -64,6 +65,16 @@ TEXT_SELECTED_OPTION = '\nSeleccionada opción {}'
 TEXT_OPTION_NOT_RECOGNIZED = 'Opción no reconocida.'
 
 TEXT_POKEMON_SEARCH_ERROR = '\nNingún Pokémon cumple con los criterios de búsqueda.'
+
+TEXT_USER = '\nUsuario: {}'
+
+TEXT_GAME = '\nPartida: {}'
+
+TEXT_NO_USERS = 'No hay usuarios registrados.'
+
+TEXT_NO_GAMEMODES = 'No hay modos de juego registrados.'
+
+TEXT_NO_GAMESESSIONS = 'No hay partidas registradas.'
 
 
 def clear():
@@ -124,7 +135,7 @@ def open_menu_gamemodes():
             gamenames_list = response.json()
 
             if not gamenames_list:
-                print('No hay modos de juego registrados.')
+                print(TEXT_NO_GAMEMODES)
                 continue
 
             print('\nModos de juego:')
@@ -136,6 +147,9 @@ def open_menu_gamemodes():
                 
             print('\nEscribe un nombre para el nuevo modo de juego:')
             name = input()
+
+            if len(name)==0:
+                continue
 
             response = requests.get(API_URL_GAMEMODE)
             gamenames_list = response.json()
@@ -186,9 +200,6 @@ def open_menu_gamemodes():
                 print('\n¿Incluir pesos pesados? Escribe 1 para sí, escribe otra cosa para no.')
                 options['powerhouse'] = input()=='1'
 
-                # print('\n¿Incluir los demás Pokémon que no pertenezcan a estas categorías? Escribe 1 para sí, escribe otra cosa para no.')
-                # options['others'] = input()=='1'
-
                 print('\nFiltrar por etapa evolutiva del Pokémon.')
                 print('¿Incluir solamente Pokémon en su última etapa evolutiva? Escribe 1 para sí, escribe otra cosa para no.')
                 options['fully_evolved'] = input()=='1'
@@ -205,7 +216,7 @@ def open_menu_gamemodes():
             gamenames_list = response.json()
 
             if not gamenames_list:
-                print('No hay modos de juego registrados.')
+                print(TEXT_NO_GAMEMODES)
                 continue
 
             print('\nModos de juego:')
@@ -215,6 +226,11 @@ def open_menu_gamemodes():
             print('\nEscribe el nombre del modo de juego que quieres eliminar:')
             name = input()
 
+            clear()
+
+            if len(name)==0:
+                continue
+
             response = requests.delete(API_URL_GAMEMODE+name)
             print(response.text)
 
@@ -223,14 +239,279 @@ def open_menu_gamemodes():
             clear()
             return
         else:
-            print('Opción no reconocida.')
-
+            print(TEXT_OPTION_NOT_RECOGNIZED)
 
 
 # MENU MANAGE USERS
 # =======================================================================
 
-def open_menu_users(self):
+def open_menu_users():
+    
+    clear()
+    
+    response = requests.get(API_URL_USER)
+    usernames_list = response.json()
+
+    while(True):
+        print('\nUsuarios:')
+        for x in usernames_list:
+            print('- {}'.format(x))
+
+        print(TEXT_MENU_MANAGE_USERS)
+
+        print(TEXT_INSERT_NUMBER)
+        option = input()
+
+        print(TEXT_SELECTED_OPTION.format(option))
+
+        #- 1: Cargar usuario
+        if(option=='1'):
+            
+            if len(usernames_list)==0:
+                print(TEXT_NO_USERS)
+                continue
+
+            while(True):
+                print('\nEscribe el nombre del usuario que quieres cargar:')
+                name = input()
+
+                if len(name)==0:
+                    clear()
+                    break
+
+                if name not in usernames_list:
+                    print('Usuario no encontrado.')
+                    continue
+
+                open_menu_gamesessions(name)
+                break
+
+        #- 2: Crear nuevo usuario
+        elif(option=='2'):
+            print('\nEscribe un nombre para el nuevo usuario:')
+            name = input()
+
+            clear()
+
+            if len(name)==0:
+                continue
+            
+            response = requests.post(API_URL_USER+name)
+            print(response.text)
+
+            # update usernames_list
+            response = requests.get(API_URL_USER)
+            usernames_list = response.json()
+
+        #- 3: Eliminar usuario
+        elif(option=='3'):
+            
+            if len(usernames_list)==0:
+                print(TEXT_NO_USERS)
+                continue
+            
+            print('\nEscribe el nombre del usuario que quieres eliminar:')
+
+            name = input()
+
+            clear()
+
+            if len(name)==0:
+                continue
+            
+            response = requests.delete(API_URL_USER+name)
+            print(response.text)
+
+            # update usernames_list
+            response = requests.get(API_URL_USER)
+            usernames_list = response.json()
+
+        #- 0: Salir
+        elif(option=='0'):
+            clear()
+            return
+        else:
+            print(TEXT_OPTION_NOT_RECOGNIZED)
+
+
+# MENU GAMESESSIONS
+# =======================================================================
+
+def open_menu_gamesessions(username:str):
+    
+    clear()
+
+    response = requests.get(API_URL_USER+username)
+    user:dict = response.json()
+    gamenames_list = user.get('games')
+
+    while(True):
+        print(TEXT_USER.format(username))
+
+        print('\nPartidas:')
+        for x in gamenames_list:
+            print('- {}'.format(x))
+
+        print(TEXT_MENU_MANAGE_GAMESESSIONS)
+        
+        print(TEXT_INSERT_NUMBER)
+        option = input()
+
+        print(TEXT_SELECTED_OPTION.format(option))
+
+        #- 1: Cargar partida
+        if(option=='1'):
+            
+            if len(gamenames_list)==0:
+                print(TEXT_NO_GAMESESSIONS)
+                continue
+
+            while(True):
+                print('\nEscribe el nombre de la partida que quieres cargar:')
+                name = input()
+
+                if len(name)==0:
+                    clear()
+                    break
+
+                if name not in gamenames_list:
+                    print('Partida no encontrada.')
+                    continue
+
+                open_menu_game(username, name)
+                break
+
+        #- 2: Apuntarse a partida
+        elif(option=='2'):
+            
+            response = requests.get(API_URL_GAMEMODE)
+            gamenames_availables_list = response.json()
+
+            if not gamenames_availables_list:
+                clear()
+                print(TEXT_NO_GAMEMODES)
+                continue
+            
+            print('\nModos de juego:')
+            for x in gamenames_availables_list:
+                print('- {}'.format(x))
+            
+            print('\nEscribe el nombre del modo de juego con el que quieres empezar una partida (creará una partida con ese mismo nombre):')
+            name = input()
+
+            clear()
+
+            if len(name)==0:
+                continue
+            
+            if name not in gamenames_availables_list:
+                print('Modo de juego no encontrado.')
+                continue
+            
+            if name in gamenames_list:
+                print('Esa partida ya fue creada anteriormente.')
+                continue
+            
+            response = requests.post(API_URL_USER_GAME.format(username, name))
+            print(response.text)
+
+            # update
+            response = requests.get(API_URL_USER+username)
+            user:dict = response.json()
+            gamenames_list = user.get('games')
+
+        #- 3: Quitarse de partida
+        elif(option=='3'):
+            
+            response = requests.get(API_URL_USER+username)
+            user:dict = response.json()
+
+            gamenames_list = user.get('games')
+
+            if not gamenames_list:
+                print(TEXT_NO_GAMESESSIONS)
+                continue
+            
+            print('\nEscribe el nombre de la partida que quieres eliminar:')
+
+            name = input()
+
+            clear()
+
+            if len(name)==0:
+                continue
+            
+            response = requests.delete(API_URL_USER_GAME.format(username, name))
+            print(response.text)
+
+            # update
+            response = requests.get(API_URL_USER+username)
+            user:dict = response.json()
+            gamenames_list = user.get('games')
+
+        #- 0: Salir
+        elif(option=='0'):
+            clear()
+            return
+        else:
+            print(TEXT_OPTION_NOT_RECOGNIZED)
+
+
+# MENU GAME
+# =======================================================================
+
+def open_menu_game(username:str, gamename:str):
+    
+    clear()
+
+    while(True):
+        print(TEXT_USER.format(username))
+        print(TEXT_GAME.format(gamename))
+        print_game_info()
+        print_box()
+
+        print(TEXT_MENU_PLAY_GAME)
+
+        print(TEXT_INSERT_NUMBER)
+        option = input()
+        clear()
+
+        print(TEXT_SELECTED_OPTION.format(option))
+
+        if(option=='1'):
+            roll()
+        elif(option=='2'):
+            roll(spend_ticket=True)
+        elif(option=='3'):
+            open_menu_cards()
+        elif(option=='4'):
+            print_filters()
+
+        elif(option=='9'):
+            clear()
+            quit()
+        elif(option=='0'):
+            clear()
+            return
+
+        else:
+            print(TEXT_OPTION_NOT_RECOGNIZED)
+
+        print(TEXT_LINE)
+
+def print_game_info():
+    pass
+
+def print_filters():
+    pass
+
+def print_box():
+    pass
+
+def roll():
+    pass
+
+def open_menu_cards():
     pass
